Refactored the convert method by removing the annonymous methods with concrete classes.
Completed the usage method that displays the annotation usage information
with a maximum width of 120 characters.

Updated the pom file to include the jar pluging that will not include
classes that are within the org.valhalla.cli.examples package.

Added a README.md file to be able to get a better discreption on github.